Dependencies/CAELUS_SmartSkies/requirements.txt
attrs==21.2.0
iniconfig==1.1.1
packaging==21.0
pluggy==1.0.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
toml==0.10.2
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.6
idna==3.2
iniconfig==1.1.1
packaging==21.0
pluggy==1.0.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
requests==2.26.0
toml==0.10.2
urllib3==1.26.7
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.6
idna==3.2
iniconfig==1.1.1
packaging==21.0
pluggy==1.0.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
python-dotenv==0.19.0
requests==2.26.0
toml==0.10.2
urllib3==1.26.7
